% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_and_return_user_arguments.R
\name{catch_and_return_user_arguments}
\alias{catch_and_return_user_arguments}
\title{Function to catch and parse user arguments for the ICOS Cities programmes.}
\usage{
catch_and_return_user_arguments()
}
\value{
Named list.
}
\description{
Function to catch and parse user arguments for the ICOS Cities programmes.
}
\author{
Stuart K. Grange
}
