% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_partectors_data.R
\name{read_partectors_data}
\alias{read_partectors_data}
\title{Function to read the Naneos Partector 2 ultrafine particle monitor data files.}
\usage{
read_partectors_data(
  file,
  tz_in_file = "UTC",
  as_long = TRUE,
  variable_switch = FALSE,
  priority_variables = FALSE,
  serial_number_as_character = FALSE,
  date_calibration_as_numeric = FALSE,
  names_to = rlang::zap(),
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{file}{A vector of Naneos Partector 2 ultrafine particle monitor data
files.}

\item{tz_in_file}{What time zone are the dates represented in the data files?}

\item{as_long}{Should the data files be returned in a "long" format?}

\item{variable_switch}{Should the data files' names be switched to more 
universal versions?}

\item{priority_variables}{If \code{as_long} and \code{variable_switch} are
\code{TRUE}, should only "priority" variables be returned?}

\item{serial_number_as_character}{Should the `serial_number` variable be
returned as a character vector?}

\item{date_calibration_as_numeric}{should the `date_calibration` variable 
be returned as a numeric vector?}

\item{names_to}{An optional character vector to add the file name of the data
file to the tibble return.}

\item{verbose}{Should the function give messages?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Tibble.
}
\description{
Function to read the Naneos Partector 2 ultrafine particle monitor data files.
}
\author{
Stuart K. Grange
}
